+++ HISTORY.md
@@ -1,3 +1,9 @@
+##1.1.1
+
++ FIX reserve computation bug.
++ FIX account.entry updating, and include signerlist info.
++ ADD signerListSet method on Transaction
+
 ##1.1.0
 
 + ADD support for multi-signature